requirements.txt
flask
requests
python-dotenv
gspread
google-auth
google-auth-oauthlib
crewai
openai
gunicorn
psycopg2-binary